fix(ui): make the blog and posts fit ui
